Move algorithm construction out of Target
