+++ probe-rs/src/config/registry.rs
@@ -431,6 +431,8 @@ fn match_name_prefix(pattern: &str, name: &str) -> bool {
 
 #[cfg(test)]
 mod tests {
+    use crate::flashing::FlashAlgorithm;
+
     use super::*;
     use std::fs::File;
     type TestResult = Result<(), RegistryError>;
@@ -508,8 +510,7 @@ mod tests {
                 // Walk through the flash algorithms and cores and try to create each one.
                 for raw_flash_algo in target.flash_algorithms.iter() {
                     for core in raw_flash_algo.cores.iter() {
-                        target
-                            .initialized_flash_algorithm_by_name(&raw_flash_algo.name, core)
+                        FlashAlgorithm::assemble_from_raw_with_core(raw_flash_algo, core, &target)
                             .unwrap_or_else(|error| {
                                 panic!(
                                     "Failed to initialize flash algorithm ({}, {}, {core}): {}",
